inference/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
numpy==1.26.4
Pillow==10.4.0
# If you have GPU/CPU constraints, pin TF accordingly. CPU-only is fine:
tensorflow==2.14.0
python-multipart==0.0.9
